basic test of locale sub-path routing
