+++ frontend/pages/_app.tsx
@@ -1,24 +1,42 @@
 import "@fortawesome/fontawesome-svg-core/styles.css"
 
-import { useEffect, useReducer } from "react"
+import {
+  useEffect,
+  useReducer,
+} from "react"
 
 import AlertContext, { Alert } from "/contexts/AlertContext"
-import { Breadcrumb, BreadcrumbContext } from "/contexts/BreadcrumbContext"
-import LanguageContext from "/contexts/LanguageContext"
+import {
+  Breadcrumb,
+  BreadcrumbContext,
+} from "/contexts/BreadcrumbContext"
 import LoginStateContext from "/contexts/LoginStateContext"
-import { isAdmin, isSignedIn } from "/lib/authentication"
-import { initGA, logPageView } from "/lib/gtag"
+import {
+  isAdmin,
+  isSignedIn,
+} from "/lib/authentication"
+import {
+  initGA,
+  logPageView,
+} from "/lib/gtag"
 import withApolloClient from "/lib/with-apollo-client"
 import { fontCss } from "/src/fonts"
 import theme from "/src/theme"
 import PagesTranslations from "/translations/pages"
 import { useTranslator } from "/util/useTranslator"
 import { ConfirmProvider } from "material-ui-confirm"
-import type { AppContext, AppProps } from "next/app"
+import type {
+  AppContext,
+  AppProps,
+} from "next/app"
 import Head from "next/head"
 import { useRouter } from "next/router"
 
-import { CacheProvider, EmotionCache, Global } from "@emotion/react"
+import {
+  CacheProvider,
+  EmotionCache,
+  Global,
+} from "@emotion/react"
 import { config as fontAwesomeConfig } from "@fortawesome/fontawesome-svg-core"
 import { CssBaseline } from "@mui/material"
 import { ThemeProvider } from "@mui/material/styles"
@@ -165,9 +183,6 @@ export function MyApp({
         <ThemeProvider theme={theme}>
           <CssBaseline />
           <LoginStateContext.Provider value={state}>
-            <LanguageContext.Provider
-              value={{ language: lng, url: languageSwitchUrl, hrefUrl }}
-            >
               <ConfirmProvider>
                 <BreadcrumbContext.Provider
                   value={{
@@ -189,7 +204,6 @@ export function MyApp({
                   </AlertContext.Provider>
                 </BreadcrumbContext.Provider>
               </ConfirmProvider>
-            </LanguageContext.Provider>
           </LoginStateContext.Provider>
         </ThemeProvider>
       </CacheProvider>
@@ -222,31 +236,8 @@ function createPath(originalUrl: string) {
 MyApp.getInitialProps = async (props: AppContext) => {
   const { ctx, Component } = props
 
-  let lng = "fi"
   let asUrl = "/"
   let hrefUrl = "/"
-
-  if (typeof window !== "undefined") {
-    if (languages.includes(ctx?.asPath?.substring(1, 3) ?? "")) {
-      lng = ctx?.asPath?.substring(1, 3) ?? ""
-    }
-
-    asUrl = ctx?.asPath ?? ""
-    hrefUrl = ctx.pathname
-  } else {
-    const maybeLng = (ctx.query.lng as string) ?? "fi"
-
-    if (languages.includes(maybeLng)) {
-      lng = maybeLng
-    } else {
-      ctx?.res?.writeHead(302, { location: "/404" })
-      ctx?.res?.end()
-    }
-
-    // @ts-ignore: TODO: check what it really is
-    asUrl = ctx?.req?.originalUrl ?? ""
-    hrefUrl = ctx.pathname //.req.path
-  }
 
   let originalProps: any = {}
 
@@ -272,10 +263,6 @@ MyApp.getInitialProps = async (props: AppContext) => {
       ...originalProps.pageProps,
       signedIn,
       admin,
-      lng,
-      asUrl,
-      languageSwitchUrl: createPath(asUrl),
-      hrefUrl,
     },
   }
 }
