+++ frontend/pages/courses/[slug]/edit.tsx
