+++ frontend/pages/courses/[slug]/index.tsx
