+++ frontend/pages/courses/[slug]/manual-completions.tsx
