+++ frontend/pages/courses/[slug]/points.tsx
