+++ frontend/pages/index.tsx
@@ -1,3 +1,17 @@
-// this is a generated export stub
+import CourseAndModuleList from "/components/Home/CourseAndModuleList"
+import EmailSubscribe from "/components/Home/EmailSubscribe"
+import ExplanationHero from "/components/Home/ExplanationHero"
+import NaviCardList from "/components/Home/NaviCardList"
 
-export { default } from "/pages/[lng]/index"
+const Home = () => {
+  return (
+    <div>
+      <ExplanationHero />
+      <NaviCardList />
+      <CourseAndModuleList />
+      <EmailSubscribe />
+    </div>
+  )
+}
+
+export default Home
