+++ frontend/pages/installation/netbeans.tsx
