+++ frontend/pages/profile/points.tsx
