+++ frontend/pages/study-modules/[slug]/edit.tsx
