+++ frontend/pages/users/search/[text]/index.tsx
